Improving "adcirc build" reproducibility and auditing.
Issue 645: Originally, the build command was build and run using
variables inside of init-adcirc.sh; this change moves the build
command into a shell script that is written inside of $ADCIRCBASE.
This also takes the place of file that was `touch`d then appended
to include the paths to the patches applied.

Adds ADCIRC_BUILD_INFO environmental variable that is defined via
the ADCIRC profile that is loaded when the 'load adcirc' command
is executed in asgsh. It points to the absolute path to the JSON
file that is generated when the ADCIRC binaries are created along
with this profile.

Resolves #645.