Added abstraction to define OPENDAPPOST for use in POSTPROCESSING.
Issue 631: opendap_post.sh is turning out to be a critical piece
of the ASGS workflow that is showing some age. The change demonstrates
its the ability to define alternate opendap post scripts by defining
OPENDAPPOST in the ASGS config file.

The file, opendap_post2.sh is meant to be a alternative opendap_post.sh
that is primarily targeted for some massive refactoring. The current
state of the file in this commit, for example:

* uses HEREDOC for all `ssh` commands (but commands issued on remote
  servers is unchanged)

* updated write-ahead logging already present to use the "WAL" prefix

Moving forward, output/opendap_post2.sh will be safe to receive
radical changes while we keep the stable script, opendap_post.sh in
a pristine condition.

Relates to #631.