Created ENV variable, OPENDAPPOST
Issue 672: Created this ENV so that we can swap out different scripts
and also reference it across scripts. It's defined in platforms.sh
like QSCRIPTGEN, etc - so it can be used and set to an alternative
script in a configuration file, e.g.

This line in the config can go from something like,

  POSTPROCESS=( createMaxCSV.sh includeWind10m.sh createOPeNDAPFileList.sh opendap_post.sh )

To,

  OPENDAPPOST=opendap_post.sh
  POSTPROCESS=( createMaxCSV.sh includeWind10m.sh createOPeNDAPFileList.sh $OPENDAPPOST )

NOTE: When used outside of the config, the full path must be specified
like in monitoring/logging.sh because of the way it's already being used
via POSTPROCESS list

Resolves #672.