Adding optional build support for minimal python 3.7.
Issue 727: Python 3.7.10 may now be installed into the ASGS Shell
Environment with the command, `build python3`. It will not conflict
with python (2.7.18). Python 2 will be deprecated before the coming
hurricane season. Also, moving forward python will not be installed
automatically unless it's deemed necessary.

Note: No non-core modules are installed, but `pip3` is installed. If
there are python3 modules that should be included by default, they
will be added to this script when identified.

Resolves #727.