Remove SCRIPTDIR from save profiles, set in local ./asgsh.
Issue 806: Isolated ASGS should share profiles, but SCRIPTDIR is a
residual setting in the profile. It probably should be set by the
local ./asgsh and not included in the profile.

If there are any issues with this change, they should only creep
up for users who are leveraging multiple ASGS installations on
the same user account, same machine.

Additionally,
  * ./update-asgs --clean (passed to asgs-brew.pl) now leaves
    existing ASGS profiles and existing ADCIRC profiles in the
    directories $HOME/.asgs and $HOME/.adcirc-meta (respectively).

  * ASGS_META_DIR replaces instances of "$HOME/.asgs", which may
    be used in theory to relocate profile locations (in the future)

Related to #794 by making profiles more independent of specific
ASGS installations.

Resolves #806.