Made init-adcirc.sh dynamic based on what's in patches/ADCIRC
Issue 845: Adding officially supported ADCIRC patchsets is now
much easier. See patches/README for information. This change
creates a basis for the use of local site ADCIRC, but this is
not yet implemented. README has been updated with some general
information on adding or modifying a supported ADCIRC version.

Related to #748.
Resolves #845.