Adding "sendNotification" file tracer back opendap_post2.sh.
Issue 919: During the refactor of opendap_post2.sh, I missed
the email comms tracer, "sendNotification" - which is meant to
act as signal for the script to send an email to the addresses
specified in the configuration file. This was getting removed bec
it isn't meant to be an actual file on the filesystem.

Resolves #919.