Added way to prepend a directory to the OpenDAP path formation.
Issue 931: Defining OPENDAPADDROOT in an ASGS configuration file will
cause an additional directory to be prepended to the data path
created and used as the upload target, in output/opendap_post2.sh.
Also writes "post.opendap.addroot" to run.properties; if not set,
it gets recorded as "null".

For example, adding the following to the ASGS config fle,

  OPENDAPADDROOT=replay-test

Will ultimately make the files accessible in a url that is rooted
at 'replay-test', e.g.,

  http://chg-1.oden.tacc.utexas.edu/thredds/catalog/asgs/replay-test/2022/al05/11/TX2008/ls6.tacc.utexas.edu/TX2008-replay-ls6/nhcConsensus/catalog.html
                                                         ^^^^^^^^^^^^

If not set, the URL would normally look as follows,

  http://chg-1.oden.tacc.utexas.edu/thredds/catalog/asgs/2022/al05/11/TX2008/ls6.tacc.utexas.edu/TX2008-replay-ls6/nhcConsensus/catalog.html
                                                         ^^^^^^^^^^^^

Resolves #931.