Adding support for bleed edge development branch of ADCIRC.
Issue 980: The current state of work/cmplrflags.mk in the development
branch of ADCIRC indicates no new structural changes, only tweaks
that seem to be artifacts of development activities. This change first
reverts the file back to a known and supported state (v55.01), then
it applies the existing set of patches that are supplied for supporting
v55.01. Testing indicates this approach works and ADCIRC gets built
as expected. ADCIRC has not been run, but this will be done outside of
this particular issue.

It is expected that support for this branch will change from time to time
since it is the target of ongoing work. However, ADCIRC is typically slow
to change; and as long as cmplrflags.mk doesn't change too often then
support should also be relativel stable. And then it would be easy to
track and fix.

Resolves #980.

(cherry picked from commit aaa3548a5cad5b5060762af88f7e866a72ec78b9)
Signed-off-by: wwlwpd <46434714+wwlwpd@users.noreply.github.com>