nhc-recorder now saves CurrentStorms.json, added blind git support.
Issue 1012: This script now saves the CurrentStorms.json file for
each storm that changes in it; this may result in duplication, but
that is probably a good thing in this case. There is now blind git
support for the currently operation environment, which will `git add`
and `git commit`/`git push` to the archives repo as new files are
added and new storms are captured.

Resolves #1012.