Updating ADCIRC patchset for latest development upstream branch.
Issue-1028: This patchset allows for the latest known working SHA
of the upstream development branch to compile. It should be adjusted
from time to time. Ideally this will be the same SHA as HEAD, but
may also not be if the latest HEAD is broken for some reason.

This commit is 2 of 2 for Issue 1028.

Resolves #1028.