Refactored bin/init-adcirc.sh to support different sources.
Issue 1004: "build adcirc" command runs bin/init-adcirc.sh behind
the scenes. This change adds support for different sources including
local versions of ADCIRC (provided it can already build) and sources
that are retreived from the upstream ADCIRC Github repository and
supports zip archives. In addition to this, patchsets can be added
for both git and remote zip file based code sources. Patchsets are
not supported for "custom" source locations intentionally.

The code was also refactored to make it much easier to add support
for additional sources in the future - with or without patcheset
support.

Resolves #1004.