Fixing `load adcirc` due to lost of `ADCIRCDIR` (put it back).
Issue 1038: The changes in #1004 took renamed ADCIRCDIR to ADCIRC_WORK_DIR,
but this had effects outside of `bin/init-adcirc.sh` (which does the
building and contains the wizard). This should not have been done since
`ADCIRCDIR` is used externally, e.g., in `./asgs_main.sh`; not to mention
the most obvious place that spawned this case, in `load adcirc`. So rather
than changing `ADCIRCDIR` everwhere, it was restored. `ADCIRCDIR` preceeded
`bin/init-adcirc.sh`.

Resolves #1038.