UX improvements.
Issue 1105: Fixed issue with the new `$PS` causing the commandline to
wrap too early and without a newline character.

Issue 1106: Adds a bash completion definition of `rebuild profile`. If
a configuration file to "rebuild" is provided at the onset, then the
operator is not prompted to specify a new one.

Resolves #1105. Resolves #1106.