Added "ls6" to bin/sq
Issue 1107: "sq" is a wrapper command around the local platform's
way of querying any batch jobs in the queue that are running or are
scheduled. Lonestar6 was not added to this when first ported.

Also remove the first line, which is a header, for the "squeue" command,
which allows the script to properly determine when there are no
jobs in the queue for $USER.

Resolves #1107.