Fixed asgs-sendmail to warn on long subject line and truncate.
Issue 1114: Additional commit to this issue, a fix to recently
discovered sensitivity to subject lengths > 78. asgs-sendmail will
now warn when the subject is > 99 characters, and in an attempt
to "just work", truncates the subject and sends it anyway.

Related to #1114.
Resolves #1121 (PR).

(cherry picked from commit 8311edc84d4220070988d3a9081944d6fb080e33)
Signed-off-by: wwlwpd <46434714+wwlwpd@users.noreply.github.com>