asgs-lint now checks veerRight/veerLeft against PERCENT value.
Issue 1189: Convention requires that the PERCENT left or right as
storm is to veer, as in: veerRight100 or veerLeft100, etc; be the
same absolute value as the PERCENT.. asgs-lint already checks to
to make sure that "Left" PERCENT is a negative integer and that
"Right" is a positive integer. This change ensures that the value
of the PERCENT is also in the "veerRight$PERCENT" or "veerLeft$PERCENT".

This is a common error when ASGS is operational.

Resolves #1189.