Began cleaning up operators view of running ASGS.
Issue 1085: This commit moves much of the non-essential chatter
that was being displayed when asgs_main.sh forecasts were being
run, to the SYSLOG where it should be. This commit makes way for
a more focused and effective experience for operators and less
"real time" users of ASGS, a like.

This change also fixes an annoying issue with "move statefile"
and "delete statefile" is executed after a "run" had concluded
or had been forcibly killed via ctrl-c.

Resolves #1085.