Added quotes around exported variables in profile for Windows paths.
Issue 1296: Windows paths were causing start up issues, and it was discovered
that the environmental variables were not being quoted. Additionally, it was
discovered and fixed that the variable, "ASGS_BREW_FLAGS", was being quoted too
early in the process. This was adjusted to be treated like the others.

In addition to this, the default "$WORK" and "$SCRATCH" directories for "debian"
were adjusted so that they wer under $HOME, and not a root level directory that
would show unnecessary warnings for non-privileged users.

Resolves #1296.
Resolves #1297 (PR).