Added ourPerl as "ourperl" to `fetch`, added to default PATHs
Issue 1364: Added our fork of Nate's Perl repo to the `fetch`
command, use `fetch ourperl`. Updated default PATHs in asgs-brew
to include the directories in git/ourPerl and git/asgs-mon; it
should be okay that they are fixed like that, it's how all the
others were added. Added also to PERL5LIB, so modules can also
be used in the general environment.

Resolves #1364.