Added support for Intel-OneAPI 2024.2 on Debian/Ubuntu
Issue 1386: Added Intel OneAPI support on Debian and Ubuntu systems that
are not handled directly in ./platforms like most HPCs are. This change
is targed towards desktop and workstation users who have installed Intel
OneAPI on to their system since Intel is now providing this free of
charge.

Related to #1386