ADCIRC build support mega-update (#1388)
Issue 1386: Added Intel OneAPI support on Debian and Ubuntu systems that
are not handled directly in ./platforms like most HPCs are. This change
is targed towards desktop and workstation users who have installed Intel
OneAPI on to their system since Intel is now providing this free of
charge.

* renamed intel-llvm to intel-oneapi

Issue 1386:

Related to #1386

* Removed "v55release", use v55.02 or v55-5bc04d6.

Issue 1386: cleaning up redundant or outdated versions of ADCIRC
in order to reduce maintenance burden.

Related to #1386.

* Added support for Intel-OneAPI 2024.2 on Debian/Ubuntu

Related to #1386

* Removing another patchset related to v55release

Issue 1386:

Related to #1386

* Removing the original v55.01 patchset.

Issue 1386:

Related to #1386.

* Removing patchset for gfortran >= 10 in lieu of a new compiler "family"

Issue 1386:

Related to #1386

* Updating v55* with debian support for Intel OneAPI

Issue 1386:

Related to #1386

* Adding Debian support for Intel OneAPI in the v55.02 patchset

Issue 1386:

Related to #1386

* v56 support to the 56.0.2 release, added Intel OneAPI support on Debian

Issue 1386:

Related to #1386

* fixup! v56 support to the 56.0.2 release, added Intel OneAPI support on Debian

* Made it more clear when there is an installation directory conflict

Issue 1386:

Related to #1386

* fixup! v56 support to the 56.0.2 release, added Intel OneAPI support on Debian

* fixup! v56 support to the 56.0.2 release, added Intel OneAPI support on Debian

* fixup! v56 support to the 56.0.2 release, added Intel OneAPI support on Debian

For merge:

Resolves to #1386
Resolves to #1383
Resolves to #1361
Resolves to #1311
Related  to #1384 (may prove viable solution, if applied to all patchesets for
all supported ADCIRC versions.