Updated handler for `list meshes`
Issue 1503: the directory structure for storing configurations
for officially supported meshes changed, but this impacted the
`list meshes` command in a way that stopped the number of nodes
from being reported. This PR fixes that.

Meshes that are contained in local assests, if set, are handled
as they were before using the original method.

Resolves #1503.