Adding support for gfortran 10.1 and greater. (#1536)
Issue 1384: This change removes the impediment inherent in `work/cmplrflags.mk` that prevented `gcc` > 9 from being used, due to the reversal of a once default flag regarding function interface mismatches. Some versions of ADCIRC may not build with newer GCC, but that's an upstream problem. For example, version v53.05 will not build with GCC 13 due to some coding issues in what is modernly expected. A new family was not added, original support for GCC/gfortran simply needed to check the version and addit if needed. Additional changes were necessary for SWAN to build, platform.pl was updated in all cases.

Resolves #1384.