Fixed OpenMPI/MPI detection regression (#1609)
Issue 1601: Detection was working before the change that broke this,
but the change was an attempt to more clearly check for the different'
cases. There are currently 4 cases for which we will skip the building
of OpenMPI, It also eliminates false positives when gfortran is the compiler
but Intel's mpif90 is in PATH. Added -c flag to define default compiler, for
use with -b for more flexible commandline switching of compilers.

Resolves #1607