Added initial singularity support for ADCIRC binaries (#1601)
Issue 1588: Added the support for ADCIRC binaries according to the
following description:

There must be a Singularity container containing all regular ADCIRC
binaries (adprep, padcirc, padcswan, etc) in the PATH of the mounted
container so that the parallel model or adcprep can function on a
compute node. Locally, the the same version that is encapsulated in
the Singularity container must be build locally, as is the normal
process. This is so any ADCIRC executible on the controller or head
node that is running ASGS is found. For example, hstime, aswip, and
adcirc (to get the version using "-v") are all expected to be available
on the head node. Since Singularity is not available on the head node
like it is on the compute nodes, we must provide 2 sets of ADCIRC.
The ADCIRC that is in the container for the compute nodes, and the
ADCIRC that is available to the "asgs_main.sh" process running on the
head node. Singularity must also be designated as an option for a
specific platform either in platforms.sh or in platforms/*/init.sh

If the particular platform definition contains a setting for the exported
environmental variable, "ASGS_SINGULARITY_CMD", running "build adcirc"
will ask the user if they wish to specifi a ".sif" file, this is
the Singularity container that is somewhere accessible on the file
system that is available to every node on the cluster (including
the head node). This information gets saved in the ADCIRC's metadata
file in $SCRIPTDIR/.adcirc-meta under the exported variable,
"ADCIRC_SINGULARITY_SIF". "load adcirc" exports this variable, and
it is now available to be detected by the qscript.pl file that
generates the batch queue scripts for each step.

If qscript.pl detects "ADCIRC_SINGULARITY_SIF", then it combines
"JOBLAUNCHER" (e.g., srun ..), "ASGS_SINGULARITY_CMD" and the
path to the .sif held by "ADCIRC_SINGULARITY_SIF" to make the correct
command to run in the batch script that runs the adcprep, padcirc,
or padcswan via singularity (and the mounting of the .sif file).

Other changes in this commit:

1. fixed the way "load adcirc" was forever prepending to PATH;
now it cleans up any ADCIRC paths before adding a new one;

2. fixing #1 made it easy to provide an "unload adcirc" command
that does the opposit of "load adcirc" - after running "unload adcirc",
no ADCIRC binaries can be found via PATH

3. "ASGS_SINGULARITY_CMD" and "ADCIRC_SINGULARITY_SIF", if defined,
are written to the run.properties file along with all the other
stuff in writeProperties.sh.

Resolves #1588.